+++ src/main/java/org/pcmm/gates/ISubscriberID.java
@@ -30,8 +30,8 @@ import org.pcmm.base.IPCMMBaseObject;
 
 public interface ISubscriberID extends IPCMMBaseObject {
 	static final short LENGTH = 8;
-	static final short SNUM = 3;
-	static final short STYPE = 1;
+	static final byte SNUM = 3;
+	static final byte STYPE = 1;
 
 	/**
 	 * source IP address for the PCMM gate.
